requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.5.0
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
curlify==2.2.1
dynaconf==3.2.2
exceptiongroup==1.1.3
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.3.0
pydantic==2.3.0
pydantic_core==2.6.3
PyHamcrest==2.0.4
pytest==7.4.0
pytest-lazy-fixture==0.6.3
requests==2.31.0
structlog==23.1.0
tomli==2.0.1
typing_extensions==4.7.1
urllib3==2.0.4
